+++ examples/count.rs
@@ -13,10 +13,7 @@ fn main() {
             let mut count = 0usize;
             loop {
                 task::sleep(time::Duration::from_secs(1)).await;
-                server
-                    .send(None, Message::Text(count.to_string()))
-                    .await
-                    .ok();
+                server.send(None, count.to_string()).await.ok();
                 count += 1;
             }
         });
@@ -25,7 +22,7 @@ fn main() {
     // Print messages sent by clients
     task::block_on(async {
         server.listen("0.0.0.0:8000").await.unwrap();
-        while let Some(event) = server.next().await {
+        while let Some(event) = server.next::<Message>().await {
             println!("{:?}", event);
         }
     });